Add logging

+++ transfiles/actions.py
@@ -1,6 +1,10 @@
 import os
 import shutil
+import logging
 from transfiles.transactions import Action
+import transfiles.logsetup
+
+logger = logging.getLogger(__name__)
 
 
 class FileAction(Action):
@@ -21,32 +25,59 @@ class MoveAction(FileAction):
         tmp_name = self.file + ".bak"
         if os.path.exists(tmp_name):
             raise RuntimeError(f"Cannot rename as {tmp_name} already exists")
-        print(f"Renaming {self.file} to {tmp_name}")
+        logger.info(f"Renaming {self.file} to {tmp_name}")
         os.rename(self.file, tmp_name)
         self.done_renaming = (self.file, tmp_name)
-        print("Done renaming")
+        logger.info("Done renaming")
 
-        # should_fail = input("Simulate exception (Y/N)?")
-        # if should_fail.lower() == "y":
-        #     raise RuntimeError("Asked to fail")
-
-        print(f"Copying {tmp_name} to {self.trg_file}")
+        logger.info(f"Copying {tmp_name} to {self.trg_file}")
+        if os.path.exists(self.trg_file):
+            # otherwise shutil.copyfile() will just silently replace it
+            raise RuntimeError("File {0} already exists".format(self.trg_file))
+        os.makedirs(os.path.dirname(self.trg_file), exist_ok=True)
         shutil.copyfile(tmp_name, self.trg_file)
-        print("Done copying")
+        logger.info("Done copying")
 
     def commit(self):
         src_file, tmp_renamed_file = self.done_renaming
-        print(f"Removing {tmp_renamed_file}")
+        logger.info(f"Removing {tmp_renamed_file}")
         os.remove(tmp_renamed_file)
-        print("Done removing")
+        logger.info("Done removing")
 
     def rollback(self):
         if self.done_renaming:
             src_file, tmp_renamed_file = self.done_renaming
-            print(f"Renaming {tmp_renamed_file} to {src_file}")
+            logger.info(f"Renaming {tmp_renamed_file} to {src_file}")
             os.rename(tmp_renamed_file, src_file)
             try:
-                print(f"Removing {self.trg_file}")
+                logger.info(f"Removing {self.trg_file}")
                 os.remove(self.trg_file)
-            except:
-                pass
+            except BaseException as e:
+                logger.error(f"Error while removing {self.trg_file}", exc_info=e)
+
+
+class CopyAction(FileAction):
+    def __init__(self, src_file, trg_file):
+        FileAction.__init__(self, src_file)
+        self.trg_file = trg_file
+
+    def __repr__(self) -> str:
+        return f"CopyAction[from: {self.file}; to: {self.trg_file}]"
+
+    def pre_commit(self):
+        if os.path.exists(self.trg_file):
+            raise RuntimeError(f"Cannot copy as {self.trg_file} already exists")
+        logger.info(f"Copying {self.file} to {self.trg_file}")
+        os.makedirs(os.path.dirname(self.trg_file), exist_ok=True)
+        shutil.copyfile(self.file, self.trg_file)
+        logger.info("Done copying")
+
+    def commit(self):
+        pass
+
+    def rollback(self):
+        try:
+            logger.info(f"Removing {self.trg_file}")
+            os.remove(self.trg_file)
+        except BaseException as e:
+            logger.error(f"Error while removing {self.trg_file}", exc_info=e)
